Struct google_servicedirectory1::api::ProjectMethods

pub struct ProjectMethods<'a, C> where
    C: 'a,  { /* fields omitted */ }

A builder providing access to all methods supported on project resources. It is not used directly, but through the ServiceDirectory hub.

Example

Instantiate a resource builder

extern crate hyper;
extern crate hyper_rustls;
extern crate yup_oauth2 as oauth2;
extern crate google_servicedirectory1 as servicedirectory1;
 
use std::default::Default;
use oauth2;
use servicedirectory1::ServiceDirectory;
 
let secret: oauth2::ApplicationSecret = Default::default();
let auth = yup_oauth2::InstalledFlowAuthenticator::builder(
        secret,
        yup_oauth2::InstalledFlowReturnMethod::HTTPRedirect,
    ).build().await.unwrap();
let mut hub = ServiceDirectory::new(hyper::Client::builder().build(hyper_rustls::HttpsConnector::with_native_roots()), auth);
// Usually you wouldn't bind this to a variable, but keep calling *CallBuilders*
// like `locations_get(...)`, `locations_list(...)`, `locations_namespaces_create(...)`, `locations_namespaces_delete(...)`, `locations_namespaces_get(...)`, `locations_namespaces_get_iam_policy(...)`, `locations_namespaces_list(...)`, `locations_namespaces_patch(...)`, `locations_namespaces_services_create(...)`, `locations_namespaces_services_delete(...)`, `locations_namespaces_services_endpoints_create(...)`, `locations_namespaces_services_endpoints_delete(...)`, `locations_namespaces_services_endpoints_get(...)`, `locations_namespaces_services_endpoints_list(...)`, `locations_namespaces_services_endpoints_patch(...)`, `locations_namespaces_services_get(...)`, `locations_namespaces_services_get_iam_policy(...)`, `locations_namespaces_services_list(...)`, `locations_namespaces_services_patch(...)`, `locations_namespaces_services_resolve(...)`, `locations_namespaces_services_set_iam_policy(...)`, `locations_namespaces_services_test_iam_permissions(...)`, `locations_namespaces_set_iam_policy(...)` and `locations_namespaces_test_iam_permissions(...)`
// to build up your call.
let rb = hub.projects();

Implementations

impl<'a, C> ProjectMethods<'a, C>

pub fn locations_namespaces_services_endpoints_create(
&self,
request: Endpoint,
parent: &str
) -> ProjectLocationNamespaceServiceEndpointCreateCall<'a, C>

Create a builder to help you perform the following task:

Creates an endpoint, and returns the new endpoint.

Arguments

request - No description provided.

parent - Required. The resource name of the service that this endpoint provides.

pub fn locations_namespaces_services_endpoints_delete(
&self,
name: &str
) -> ProjectLocationNamespaceServiceEndpointDeleteCall<'a, C>

Create a builder to help you perform the following task:

Deletes an endpoint.

Arguments

name - Required. The name of the endpoint to delete.

pub fn locations_namespaces_services_endpoints_get(
&self,
name: &str
) -> ProjectLocationNamespaceServiceEndpointGetCall<'a, C>

Create a builder to help you perform the following task:

Gets an endpoint.

Arguments

name - Required. The name of the endpoint to get.

pub fn locations_namespaces_services_endpoints_list(
&self,
parent: &str
) -> ProjectLocationNamespaceServiceEndpointListCall<'a, C>

Create a builder to help you perform the following task:

Lists all endpoints.

Arguments

parent - Required. The resource name of the service whose endpoints you'd like to list.

pub fn locations_namespaces_services_endpoints_patch(
&self,
request: Endpoint,
name: &str
) -> ProjectLocationNamespaceServiceEndpointPatchCall<'a, C>

Create a builder to help you perform the following task:

Updates an endpoint.

Arguments

request - No description provided.

name - Immutable. The resource name for the endpoint in the format projects/*/locations/*/namespaces/*/services/*/endpoints/*.

pub fn locations_namespaces_services_create(
&self,
request: Service,
parent: &str
) -> ProjectLocationNamespaceServiceCreateCall<'a, C>

Create a builder to help you perform the following task:

Creates a service, and returns the new service.

Arguments

request - No description provided.

parent - Required. The resource name of the namespace this service will belong to.

pub fn locations_namespaces_services_delete(
&self,
name: &str
) -> ProjectLocationNamespaceServiceDeleteCall<'a, C>

Create a builder to help you perform the following task:

Deletes a service. This also deletes all endpoints associated with the service.

Arguments

name - Required. The name of the service to delete.

pub fn locations_namespaces_services_get(
&self,
name: &str
) -> ProjectLocationNamespaceServiceGetCall<'a, C>

Create a builder to help you perform the following task:

Gets a service.

Arguments

name - Required. The name of the service to get.

pub fn locations_namespaces_services_get_iam_policy(
&self,
request: GetIamPolicyRequest,
resource: &str
) -> ProjectLocationNamespaceServiceGetIamPolicyCall<'a, C>

Create a builder to help you perform the following task:

Gets the IAM Policy for a resource (namespace or service only).

Arguments

request - No description provided.

resource - REQUIRED: The resource for which the policy is being requested. See the operation documentation for the appropriate value for this field.

pub fn locations_namespaces_services_list(
&self,
parent: &str
) -> ProjectLocationNamespaceServiceListCall<'a, C>

Create a builder to help you perform the following task:

Lists all services belonging to a namespace.

Arguments

parent - Required. The resource name of the namespace whose services you'd like to list.

pub fn locations_namespaces_services_patch(
&self,
request: Service,
name: &str
) -> ProjectLocationNamespaceServicePatchCall<'a, C>

Create a builder to help you perform the following task:

Updates a service.

Arguments

request - No description provided.

name - Immutable. The resource name for the service in the format projects/*/locations/*/namespaces/*/services/*.

pub fn locations_namespaces_services_resolve(
&self,
request: ResolveServiceRequest,
name: &str
) -> ProjectLocationNamespaceServiceResolveCall<'a, C>

Create a builder to help you perform the following task:

Returns a service and its associated endpoints. Resolving a service is not considered an active developer method.

Arguments

request - No description provided.

name - Required. The name of the service to resolve.

pub fn locations_namespaces_services_set_iam_policy(
&self,
request: SetIamPolicyRequest,
resource: &str
) -> ProjectLocationNamespaceServiceSetIamPolicyCall<'a, C>

Create a builder to help you perform the following task:

Sets the IAM Policy for a resource (namespace or service only).

Arguments

request - No description provided.

resource - REQUIRED: The resource for which the policy is being specified. See the operation documentation for the appropriate value for this field.

pub fn locations_namespaces_services_test_iam_permissions(
&self,
request: TestIamPermissionsRequest,
resource: &str
) -> ProjectLocationNamespaceServiceTestIamPermissionCall<'a, C>

Create a builder to help you perform the following task:

Tests IAM permissions for a resource (namespace or service only).

Arguments

request - No description provided.

resource - REQUIRED: The resource for which the policy detail is being requested. See the operation documentation for the appropriate value for this field.

pub fn locations_namespaces_create(
&self,
request: Namespace,
parent: &str
) -> ProjectLocationNamespaceCreateCall<'a, C>

Create a builder to help you perform the following task:

Creates a namespace, and returns the new namespace.

Arguments

request - No description provided.

parent - Required. The resource name of the project and location the namespace will be created in.

pub fn locations_namespaces_delete(
&self,
name: &str
) -> ProjectLocationNamespaceDeleteCall<'a, C>

Create a builder to help you perform the following task:

Deletes a namespace. This also deletes all services and endpoints in the namespace.

Arguments

name - Required. The name of the namespace to delete.

pub fn locations_namespaces_get(
&self,
name: &str
) -> ProjectLocationNamespaceGetCall<'a, C>

Create a builder to help you perform the following task:

Gets a namespace.

Arguments

name - Required. The name of the namespace to retrieve.

pub fn locations_namespaces_get_iam_policy(
&self,
request: GetIamPolicyRequest,
resource: &str
) -> ProjectLocationNamespaceGetIamPolicyCall<'a, C>

Create a builder to help you perform the following task:

Gets the IAM Policy for a resource (namespace or service only).

Arguments

request - No description provided.

resource - REQUIRED: The resource for which the policy is being requested. See the operation documentation for the appropriate value for this field.

pub fn locations_namespaces_list(
&self,
parent: &str
) -> ProjectLocationNamespaceListCall<'a, C>

Create a builder to help you perform the following task:

Lists all namespaces.

Arguments

parent - Required. The resource name of the project and location whose namespaces you'd like to list.

pub fn locations_namespaces_patch(
&self,
request: Namespace,
name: &str
) -> ProjectLocationNamespacePatchCall<'a, C>

Create a builder to help you perform the following task:

Updates a namespace.

Arguments

request - No description provided.

name - Immutable. The resource name for the namespace in the format projects/*/locations/*/namespaces/*.

pub fn locations_namespaces_set_iam_policy(
&self,
request: SetIamPolicyRequest,
resource: &str
) -> ProjectLocationNamespaceSetIamPolicyCall<'a, C>

Create a builder to help you perform the following task:

Sets the IAM Policy for a resource (namespace or service only).

Arguments

request - No description provided.

resource - REQUIRED: The resource for which the policy is being specified. See the operation documentation for the appropriate value for this field.

pub fn locations_namespaces_test_iam_permissions(
&self,
request: TestIamPermissionsRequest,
resource: &str
) -> ProjectLocationNamespaceTestIamPermissionCall<'a, C>

Create a builder to help you perform the following task:

Tests IAM permissions for a resource (namespace or service only).

Arguments

request - No description provided.

resource - REQUIRED: The resource for which the policy detail is being requested. See the operation documentation for the appropriate value for this field.

pub fn locations_get(&self, name: &str) -> ProjectLocationGetCall<'a, C>

Create a builder to help you perform the following task:

Gets information about a location.

Arguments

name - Resource name for the location.

pub fn locations_list(&self, name: &str) -> ProjectLocationListCall<'a, C>

Create a builder to help you perform the following task:

Lists information about the supported locations for this service.

Arguments

name - The resource that owns the locations collection, if applicable.